Moved all msg types to common contract crate
