+++ common/mixnet-contract/src/lib.rs
@@ -1,34 +1,15 @@
-use serde::{Deserialize, Serialize};
+// Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
 
 mod delegation;
 mod gateway;
 mod mixnode;
+mod msg;
+mod types;
 
 pub use cosmwasm_std::{Addr, Coin};
 pub use delegation::{Delegation, PagedGatewayDelegationsResponse, PagedMixDelegationsResponse};
 pub use gateway::{Gateway, GatewayBond, GatewayOwnershipResponse, PagedGatewayResponse};
 pub use mixnode::{Layer, MixNode, MixNodeBond, MixOwnershipResponse, PagedMixnodeResponse};
-
-#[derive(Debug, Default, Serialize, Deserialize, Copy, Clone, Eq, PartialEq)]
-pub struct LayerDistribution {
-    pub gateways: u64,
-    pub layer1: u64,
-    pub layer2: u64,
-    pub layer3: u64,
-}
-
-impl LayerDistribution {
-    pub fn choose_with_fewest(&self) -> Layer {
-        let layers = [
-            (Layer::One, self.layer1),
-            (Layer::Two, self.layer2),
-            (Layer::Three, self.layer3),
-        ];
-        layers.iter().min_by_key(|x| x.1).unwrap().0
-    }
-}
-
-// type aliases for better reasoning about available data
-pub type IdentityKey = String;
-pub type IdentityKeyRef<'a> = &'a str;
-pub type SphinxKey = String;
+pub use msg::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg};
+pub use types::{IdentityKey, IdentityKeyRef, LayerDistribution, SphinxKey, StateParams};
