+++ contracts/mixnet/src/lib.rs
@@ -1,7 +1,9 @@
+// Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
 pub mod contract;
 pub mod error;
 pub(crate) mod helpers;
-pub mod msg;
 pub mod queries;
 pub mod state;
 pub(crate) mod storage;
